% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{append}
\alias{append}
\title{Append to a vector}
\usage{
append(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{object}
}
\value{
NULL. Modifies the input vector.
}
\description{
Append to a vector
}
\examples{
  x <- vector()
  for (i in 1:10) append(x, i)
  x
}
