% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{check_comprehension}
\alias{check_comprehension}
\title{Check comprehension is well formed}
\usage{
check_comprehension(expr)
}
\arguments{
\item{expr}{a call to comprehension}
}
\value{
NULL. Throws an error if there's an issue.
}
\description{
Check comprehension is well formed
}
