% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehension.R
\name{comprehend}
\alias{comprehend}
\title{Comprehension}
\usage{
comprehend(comprehension_call, type)
}
\arguments{
\item{comprehension_call}{expression, a comprehension call}

\item{type}{character, giving the type of the output}
}
\value{
A vector of mode `type`.
}
\description{
Comprehension
}
