% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_utilities.R
\name{conjunction}
\alias{conjunction}
\title{Join expressions with `&`}
\usage{
conjunction(exprs)
}
\arguments{
\item{exprs}{a list of expressions}
}
\value{
An expression.
}
\description{
Join expressions with `&`
}
