% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_utilities.R
\name{construct_loop}
\alias{construct_loop}
\title{Construct a loop call}
\usage{
construct_loop(range, filters, body)
}
\arguments{
\item{range}{expression of the form `<variable> %in% <range>`}

\item{filters}{a list of expressions}

\item{body}{an expression}
}
\value{
A loop expression.
}
\description{
Construct a loop call
}
