% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_utilities.R
\name{contains}
\alias{contains}
\title{Test if expression contain a symbol}
\usage{
contains(expr, sym)
}
\arguments{
\item{expr}{an expression}

\item{sym}{a symbol}
}
\value{
Boolean.
}
\description{
Test if expression contain a symbol
}
