% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehension.R
\name{.}
\alias{.}
\title{Comprehension}
\usage{
.(comprehension, ..., type = "list")
}
\arguments{
\item{comprehension}{a comprehension expression of the form
  `<calculation> | <range 1> [& <range2> ...]`
where each `<range>` has the form
  `<variable> %in% <vector>`}

\item{...}{filter conditions}

\item{type}{character, giving the type of the output vector}
}
\value{
A vector of mode `type`, possibly coerced to a more general type.
}
\description{
A Python like list comprehension; `.` has a type argument for the output
  (with default value "list"), `l` returns a list and `v` returns a
  homogeneous vector.
}
\examples{
  .(x^2 | x \%in\% 1:10)
  .(x^2 | x \%in\% 1:10, x \%\% 3 == 0, type = "numeric")
  .(c(x, y, z) | x \%in\% 1:20 & y \%in\% x:20 & z \%in\% y:20,
    x^2 + y^2 == z^2)
  .(x + y + z | c(x, y, z) \%in\% zip(1:10, 10:1, c(-5:-1, 1:5)))

  l(c(x, y, z) | x \%in\% 1:20 & y \%in\% x:20 & z \%in\% y:20,
    x^2 + y^2 == z^2)
  v(x^2 | x \%in\% 1:10, x \%\% 3 == 0)
}
