% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_utilities.R
\name{drop}
\alias{drop}
\title{Remove named argument from call}
\usage{
drop(expr, x)
}
\arguments{
\item{expr}{a call expression}

\item{x}{character}
}
\value{
Call expression `expr` without argument `x`.
}
\description{
Remove named argument from call
}
