% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehension.R
\name{expand}
\alias{expand}
\title{Expand comprehension variables}
\usage{
expand(vars)
}
\arguments{
\item{vars}{"vars" component of output of `parse_comprehension`}
}
\value{
a character vector
}
\description{
Expand comprehension variables
}
