% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehension.R
\name{loop}
\alias{loop}
\title{Construct comprehension loop}
\usage{
loop(components, var, body)
}
\arguments{
\item{components}{output of `parse_comprehension`}

\item{var}{character, giving name of one comprehension variable}

\item{body}{expression, body of the comprehension}
}
\value{
A loop expression.
}
\description{
Construct comprehension loop
}
