% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehension.R
\name{make_comprehension_code}
\alias{make_comprehension_code}
\title{Convert comprehension call into loop code}
\usage{
make_comprehension_code(components)
}
\arguments{
\item{components}{output of `parse_comprehension`}
}
\value{
An expression.
}
\description{
Convert comprehension call into loop code
}
