% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{organise_filters}
\alias{organise_filters}
\title{Group filter conditions by when they can be applied}
\usage{
organise_filters(filters, vars)
}
\arguments{
\item{filters}{list of expressions}

\item{vars}{character vector of variables used in filters}
}
\value{
A named list with filters that can only be applied once the named
variable is introduced.
}
\description{
Group filter conditions by when they can be applied
}
