% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_comprehension}
\alias{parse_comprehension}
\title{Parse a comprehension call}
\usage{
parse_comprehension(call)
}
\arguments{
\item{call}{expression, giving a comprehension call}
}
\value{
List with values calculation, vars, ranges, and filters; vars is a
  character vector, the other values are expressions (calculation) or lists
  of expressions (ranges, filters); elements of the ranges list have the form
  `<variable> %in% <range>`.
}
\description{
Parse a comprehension call
}
