% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_utilities.R
\name{separate_into_conjuncts}
\alias{separate_into_conjuncts}
\title{Split expression on `&`}
\usage{
separate_into_conjuncts(expr)
}
\arguments{
\item{expr}{an expression}
}
\value{
A list of expressions.
}
\description{
Split expression on `&`
}
