% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehension.R
\name{split}
\alias{split}
\title{Deal with joined variables}
\usage{
split(loop_expr, new_var)
}
\arguments{
\item{loop_expr}{a loop expression from `loop`}

\item{new_var}{character, giving name for new loop variable}
}
\value{
A loop expression.
}
\description{
Converts a comprehension loop with range over more than one variable (like
  `c(x, y) %in% zip(a, b)`) into one over a single variable.
}
